parser/requirements.txt
Wikipedia-API==0.5.4
python-slugify==6.1.1
tqdm==4.62.3